+++ packages/dropzone/src/components/dropzone/index.ts
@@ -1,0 +1,4 @@
+export * from './FileIcon';
+export * from './FilePreview';
+export * from './Dropzone';
+export * from './FilePreview';